pyproject.toml
[project]
name = "caring-note-prompt-tuner"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asyncpg>=0.30.0",
    "ipykernel>=6.29.5",
    "jupyterlab>=4.4.1",
    "matplotlib>=3.10.1",
    "notebook>=7.4.1",
    "openai-agents>=0.0.14",
    "pandas>=2.2.3",
    "psycopg2-binary>=2.9.10",
    "pydantic>=2.11.4",
    "python-dotenv>=1.1.0",
    "sqlalchemy>=2.0.40",
    "sqlmodel>=0.0.24",
]
